requirements.txt
# Core
fastapi>=0.111.0
uvicorn>=0.27.0
python-dotenv>=1.0.0

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Email
email-validator>=2.0.0

# Gmail API
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.17.3

aiofiles>=22.1.0
